Refactor MemberSection
