+++ src/containers/homepage/style.module.scss
@@ -74,19 +74,3 @@ $max-width: 800px;
     pointer-events: none;
   }
 }
-
-.member {
-  width: 12.5%;
-  min-width: 80px;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  margin: 0 auto;
-  padding: 0.25rem;
-  text-align: center;
-
-  img {
-    max-width: 100%;
-    margin-bottom: 0.5rem;
-  }
-}
